+++ src/mrpro/data/KHeader.py
@@ -1,7 +1,9 @@
 """MR raw data / k-space data header dataclass."""
 
+from __future__ import annotations
+
 import dataclasses
-import datetime as dt
+import datetime
 import warnings
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING
@@ -14,7 +16,6 @@ from mrpro.data import enums
 from mrpro.data.AcqInfo import AcqInfo
 from mrpro.data.MoveDataMixin import MoveDataMixin
 from mrpro.data.SpatialDimension import SpatialDimension
-from mrpro.data.traj_calculators.KTrajectoryCalculator import KTrajectoryCalculator
 from mrpro.utils.summarize_tensorvalues import summarize_tensorvalues
 from mrpro.utils.unit_conversion import deg_to_rad, mm_to_m, ms_to_s
 
@@ -56,7 +57,7 @@ class KHeader(MoveDataMixin):
     lamor_frequency_proton: float | None = None
     """Lamor frequency of hydrogen nuclei [Hz]."""
 
-    datetime: dt.datetime | None = None
+    datetime: datetime.datetime | None = None
     """Date and time of acquisition."""
 
     te: list[float] | torch.Tensor = field(default_factory=list)
@@ -188,13 +189,13 @@ class KHeader(MoveDataMixin):
         elif header.studyInformation is not None and header.studyInformation.studyTime is not None:
             time = header.studyInformation.studyTime.to_time()
         else:  # if no time is given, set to 00:00:00
-            time = dt.time()
+            time = datetime.time()
         if header.measurementInformation is not None and header.measurementInformation.seriesDate is not None:
             date = header.measurementInformation.seriesDate.to_date()
-            parameters['datetime'] = dt.datetime.combine(date, time)
+            parameters['datetime'] = datetime.datetime.combine(date, time)
         elif header.studyInformation is not None and header.studyInformation.studyDate is not None:
             date = header.studyInformation.studyDate.to_date()
-            parameters['datetime'] = dt.datetime.combine(date, time)
+            parameters['datetime'] = datetime.datetime.combine(date, time)
 
         if header.subjectInformation is not None and header.subjectInformation.patientName is not None:
             parameters['patient_name'] = header.subjectInformation.patientName
