+++ src/mrpro/data/__init__.py
@@ -17,8 +17,6 @@ from mrpro.data.QData import QData
 from mrpro.data.QHeader import QHeader
 from mrpro.data.Rotation import Rotation
 from mrpro.data.SpatialDimension import SpatialDimension
-from mrpro.data.ReduceRepeatMixin import ReduceRepeatMixin
-
 __all__ = [
     "AcqIdx",
     "AcqInfo",
@@ -38,7 +36,6 @@ __all__ = [
     "MoveDataMixin",
     "QData",
     "QHeader",
-    "ReduceRepeatMixin",
     "Rotation",
     "SpatialDimension",
     "acq_filters",
